+++ tests/integration/models/reddit/test_redditor.py
@@ -28,7 +28,7 @@ class TestRedditor(IntegrationTest):
         with pytest.raises(RedditAPIException) as excinfo:
             redditor = await reddit.redditor(self.FRIEND.lower(), fetch=True)
             await redditor.friend(note="asyncpraw")
-        assert "GOLD_REQUIRED" == excinfo.value.error_type
+        assert "GOLD_REQUIRED" == excinfo.value.items[0].error_type
 
     async def test_friend_info(self, reddit):
         reddit.read_only = False
@@ -50,7 +50,7 @@ class TestRedditor(IntegrationTest):
         with pytest.raises(RedditAPIException) as excinfo:
             redditor = await reddit.redditor("subreddit_stats")
             await redditor.gild()
-        assert "INSUFFICIENT_COINS" == excinfo.value.error_type
+        assert "INSUFFICIENT_COINS" == excinfo.value.items[0].error_type
 
     async def test_message(self, reddit):
         reddit.read_only = False
@@ -120,7 +120,7 @@ class TestRedditor(IntegrationTest):
         redditor = Redditor(reddit, "thisusershouldnotexist")
         with pytest.raises(RedditAPIException) as excinfo:
             await redditor.trophies()
-        assert "USER_DOESNT_EXIST" == excinfo.value.error_type
+        assert "USER_DOESNT_EXIST" == excinfo.value.items[0].error_type
 
     async def test_unblock(self, reddit):
         reddit.read_only = False
@@ -282,7 +282,7 @@ class TestRedditorListings(IntegrationTest):
         await redditor.block()
         with pytest.raises(RedditAPIException) as excinfo:
             await redditor.trust()
-        assert "CANT_WHITELIST_AN_ENEMY" == excinfo.value.error_type
+        assert "CANT_WHITELIST_AN_ENEMY" == excinfo.value.items[0].error_type
 
     async def test_upvoted(self, reddit):
         reddit.read_only = False